% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_colsa.R
\docType{data}
\name{covid_colsa}
\alias{covid_colsa}
\title{Covid-19 Data for Colombia & South Africa}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 542 rows and 35 columns.
}
\source{
The data comes from Guidotti, E., Ardia, D., (2020), "COVID-19 Data Hub", Journal of Open Source Software 5(51):2376, doi: 10.21105/joss.02376, through the R package "COVID19" and filtered for the country Colombia and South Africa.
}
\usage{
covid_colsa
}
\arguments{
\item{id}{Country Code}

\item{date}{Date of observation}

\item{tests}{Count of tests for a date}

\item{confirmed}{Count of confirmed for a date}

\item{recovered}{Count of recovered for a date}

\item{deaths}{Count of deaths for a date}

\item{key_apple_mobility}{Country Name}
}
\description{
This data contains 35 variables for Covid stats in Colombia and South Africa from January 2020.
}
\keyword{datasets}
