% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyplot.R
\name{dailyplot}
\alias{dailyplot}
\title{Daily Count Plot}
\usage{
dailyplot(data)
}
\arguments{
\item{data}{A dataset which consists of daily counts of all attributes.}
}
\description{
This function automatically launches the Shiny App and makes it available to the user to have an overview of how the datasets are used to compare the stats in Colombia and South Africa. This App can be used as a starting point to learn the potential visualizations using the data available in 'covidcolsa'.
}
