% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the Shiny App}
\usage{
launch_app()
}
\description{
This function automatically launches the Shiny App and makes it available to the user to have an overview of how the datasets are used to compare the stats in Colombia and South Africa. This App can be used as a starting point to learn the potential visualizations using the data available in 'covidcolsa'.
@export
}
